fixed another tween issue -608 to -508

+++ nova2d/nova-zero/src/graphics/Text.cpp
@@ -111,8 +111,8 @@ namespace novazero
 			}
 			else
 			{
-				m_DrawRect.x = oX + OffsetX();
-				m_DrawRect.y = oY + OffsetY();
+				m_DrawRect.x = (int)(oX + OffsetX());
+				m_DrawRect.y = (int)(oY + OffsetY());
 			}	
 
 			int w = m_DrawRect.w;
